--- src/main/java/com/example/project/Tile.java
@@ -1,31 +1,0 @@
-package com.example.project;
-
-public class Tile {
-
-    private final char letter;
-
-    private final int value;
-
-
-    public Tile(char letter, int value) {
-
-        this.letter = letter;
-
-        this.value = value;
-
-    }
-
-
-    public char getLetter(){
-
-        return letter;
-
-    }
-
-    public int getValue(){
-
-        return value;
-
-    }
-
-}
